Added LinkedList implementation that conforms to ListInterface. Updated indexing operations to use size_t as the type of an index. Fixed Unix directory operations. DSpec compiles.

+++ tui/filebox.d
@@ -276,7 +276,7 @@ class TuiFileBox : TuiWidget, ListInterface!(String) {
 		return _list.opApply(loopFunc);
 	}
 	
-	int opApply(int delegate(ref int, ref String) loopFunc) {
+	int opApply(int delegate(ref size_t, ref String) loopFunc) {
 		return _list.opApply(loopFunc);
 	}
